Updated TicketPage.jsx
